# docs/swagger.yaml
...
basePath: /
definitions:
...
  entity.Site:
    properties:
      base_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      icon_url:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
  handler.RespBody:
    properties:
      code:
        description: http code
        type: integer
      data:
        description: response data
      msg:
        description: response message
        type: string
      reason:
        description: reason key
        type: string
    type: object
...
  schema.AdminProfileTagInfo:
    properties:
      description:
        type: string
      id:
        type: string
      kind:
        type: integer
      name:
        type: string
      slug:
        type: string
      status:
        type: integer
    type: object
  schema.AdminProfileTagUpsertReq:
    properties:
      description:
        maxLength: 512
        type: string
      kind:
        enum:
        - 1
        - 2
        - 3
        type: integer
      name:
        maxLength: 128
        minLength: 1
        type: string
      slug:
        maxLength: 64
        minLength: 2
        type: string
      status:
        enum:
        - 1
        - 9
        type: integer
    required:
    - name
    - slug
    type: object
...
  schema.ProfileTagInfo:
    properties:
      description:
        type: string
      id:
        type: string
      kind:
        type: integer
      name:
        type: string
      slug:
        type: string
    type: object
...
  schema.SiteAddReq:
    properties:
      base_url:
        type: string
      description:
        type: string
      name:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  schema.SiteStatusReq:
    properties:
      active:
        type: boolean
      site_id:
        type: string
    required:
    - site_id
    type: object
...
  schema.SiteUserRoleReq:
    properties:
      role_id:
        type: integer
      site_id:
        type: string
      user_id:
        type: string
    required:
    - role_id
    - site_id
    - user_id
    type: object
...
  site.SiteUserRole:
    properties:
      role_id:
        type: integer
      site_id:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: Apache Answer API
  title: Apache Answer
paths:
...
  /answer/admin/api/network/tags:
    get:
      description: list the profile tag catalog including inactive entries
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/schema.AdminProfileTagInfo'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: list all profile tags
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: add a curated tag to the profile tag catalog
      parameters:
      - description: tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AdminProfileTagUpsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.ProfileTagInfo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: create a profile tag
      tags:
      - admin
  /answer/admin/api/network/tags/{id}:
    put:
      consumes:
      - application/json
      description: update a profile tag's name, kind, description or status
      parameters:
      - description: tag id
        in: path
        name: id
        required: true
        type: string
      - description: tag
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AdminProfileTagUpsertReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/schema.ProfileTagInfo'
              type: object
      security:
      - ApiKeyAuth: []
      summary: update a profile tag
      tags:
      - admin
...
  /answer/admin/api/site:
    get:
      description: get a sub-site by id
      parameters:
      - description: site id
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/entity.Site'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get a sub-site by id
      tags:
      - admin
    post:
      consumes:
      - application/json
      description: create a sub-site
      parameters:
      - description: site
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteAddReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/entity.Site'
              type: object
      security:
      - ApiKeyAuth: []
      summary: create a sub-site
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: update a sub-site; empty status/slug fields leave the stored values
        untouched
      parameters:
      - description: site
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.Site'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: update a sub-site
      tags:
      - admin
  /answer/admin/api/site/role:
    get:
      description: get a user's effective role on a sub-site
      parameters:
      - description: user id
        in: query
        name: user_id
        required: true
        type: string
      - description: site id
        in: query
        name: site_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  $ref: '#/definitions/site.SiteUserRole'
              type: object
      security:
      - ApiKeyAuth: []
      summary: get a user's role on a sub-site
      tags:
      - admin
    put:
      consumes:
      - application/json
      description: assign a user's role on a sub-site; a site role can raise but never
        lower the global role
      parameters:
      - description: role assignment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteUserRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: assign a user's role on a sub-site
      tags:
      - admin
  /answer/admin/api/site/status:
    put:
      consumes:
      - application/json
      description: activate or suspend a sub-site; the default site cannot be suspended
      parameters:
      - description: status
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.SiteStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.RespBody'
      security:
      - ApiKeyAuth: []
      summary: activate or suspend a sub-site
      tags:
      - admin
...
  /answer/admin/api/sites:
    get:
      description: list all sub-sites, including suspended ones
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.RespBody'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/entity.Site'
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: list all sub-sites
      tags:
      - admin
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"