# internal/schema/site_schema.go
/*
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

package schema

// SiteAddReq creates a sub-site.
type SiteAddReq struct {
	Name        string `json:"name" binding:"required"`
	Slug        string `json:"slug" binding:"required"`
	Description string `json:"description"`
	BaseURL     string `json:"base_url"`
}

// SiteStatusReq activates or suspends a sub-site.
type SiteStatusReq struct {
	SiteID string `json:"site_id" binding:"required"`
	Active bool   `json:"active"`
}

// SiteUserRoleReq assigns a user's role on one sub-site.
type SiteUserRoleReq struct {
	UserID string `json:"user_id" binding:"required"`
	SiteID string `json:"site_id" binding:"required"`
	RoleID int    `json:"role_id" binding:"required"`
}